audit fix F-001b/c (money-MOVING §26 contract, USER-signed-off): BigInt-safe operator wire
Make the operator seamless-wallet wire money BigInt-safe. The RGS->operator `amount` and the
operator->RGS `balance` are now decimal STRINGS of minor units (were JSON numbers) — a high-decimal
currency value (>= 2^53 minor units, e.g. ETH wei) silently corrupted through Number()/Math.trunc.

Contract (operator-wallet.types.ts): OperatorBetRequest/WinRequest.amount: string;
OperatorWalletResponse.balance: string; OperatorWalletApi.balance(): Promise<string>.
seamless-operator-wallet: move(amount: bigint) sends amount.toString(); getBalance/move return
BigInt(res.balance) (dropped Math.trunc). http-operator-wallet: balance()/move() return res.balance
as a string (dropped Math.trunc). MockOperator: bigint-internal, BigInt(req.amount), string balances,
balanceOf(): bigint.

PURE ENCODING change — no money value/logic/idempotency altered (win-never-rolls-back H2,
bet-compensates-on-timeout, insufficient-clean-reject, transactionId dedup, rollback/reverse all
unchanged); also fails SAFE on a malformed operator balance (BigInt throws loudly vs Math.trunc
silent-floor). Verified GO by money-path-auditor.

Docs: api-integration-spec §7/§9 + Postman folder-5 operator amount/balance -> strings (the
published operator contract). Player-facing serialization (F-019) intentionally LEFT as number
(separate, deferred finding).

Tests: 15 operator test files updated number->bigint/string (no assertion weakened, no magnitude
changed) + the MockOperator/sandbox-operator helpers + 2 high-decimal e2e (wei-scale, guard-the-guard:
the value provably corrupts through a JS number, exact through the string wire). Full suite
949 pass / 3 skip / 0 fail; tsc (src+test) 0.

FOLLOW-UP (F-078, soak/sandbox tooling, NOT in this commit): align load/operator-wallet-stub.ts +
scripts/operator-recon-check.ts + the /debug/book dump to the string/bigint contract (works at the
current low-value sandbox scale; needed for the deferred high-dp operator soak).

+++ test/ha/cashout-authority.test.ts
@@ -163,7 +163,7 @@ describe("cashOut authoritative crash gate (Finding 1) — DB round is the sourc
     // ACTIVE bet on a betting round.
     activeRoundId = await makeRound("betting", 5.0);
     const betId = await placeActiveBet(userId, walletId, "A", stake);
-    expect(operatorApi.balanceOf(playerId, currency)).toBe(startBalance - stake);
+    expect(operatorApi.balanceOf(playerId, currency)).toBe(BigInt(startBalance - stake));
     const winsBefore = operatorApi.calls.win;
 
     // Engine running + DB round AUTHORITATIVELY running, crashPoint 5.0 > mult 2.0 → legit.
@@ -182,7 +182,7 @@ describe("cashOut authoritative crash gate (Finding 1) — DB round is the sourc
     expect(after.payoutTxId).toBeTruthy();
 
     // Operator was credited exactly once: 100_000 − 1000 + 2000.
-    expect(operatorApi.balanceOf(playerId, currency)).toBe(startBalance - stake + stake * 2);
+    expect(operatorApi.balanceOf(playerId, currency)).toBe(BigInt(startBalance - stake + stake * 2));
     expect(operatorApi.calls.win).toBe(winsBefore + 1);
   });
 
@@ -196,8 +196,8 @@ describe("cashOut authoritative crash gate (Finding 1) — DB round is the sourc
 
     activeRoundId = await makeRound("betting", 5.0);
     const betId = await placeActiveBet(userId, walletId, "A", stake);
-    const balAfterStake = operatorApi.balanceOf(playerId, currency); // 99_000
-    expect(balAfterStake).toBe(startBalance - stake);
+    const balAfterStake = operatorApi.balanceOf(playerId, currency); // 99_000n
+    expect(balAfterStake).toBe(BigInt(startBalance - stake));
     const winsBefore = operatorApi.calls.win;
     const creditRowsBefore = await payoutCreditRows(walletId);
